requirements.txt
langgraph
typing
langchain
langchain-core
langchain-openai
load_dotenv
streamlit
langgraph-checkpoint-sqlite
langchain_community
ddgs